requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
scikit-learn
seaborn
matplotlib
jupyterlab
